+++ src/mastodon_api/timelines/types.rs
@@ -2,8 +2,8 @@ use serde::Deserialize;
 use uuid::Uuid;
 
 use crate::mastodon_api::{
-    deserializers::deserialize_boolean,
     pagination::PageSize,
+    serializers::deserialize_boolean,
 };
 
 fn default_page_size() -> PageSize { PageSize::new(20) }