+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed incorrect processing of poll options with HTML special characters.
+- Use correct datetime format in Mastodon API responses.
 
 ## [3.17.0] - 2025-02-13